+++ rocketgram/bot.py
@@ -10,7 +10,9 @@ from contextlib import suppress
 
 from .errors import RocketgramRequest429Error, RocketgramStopRequest
 from .errors import RocketgramRequestError, RocketgramRequest400Error, RocketgramRequest401Error
-from .requests import *
+from .context import context2
+from .requests import Request
+from typing import TYPE_CHECKING, List, Optional
 from .update import Update, UpdateType, Response
 
 if TYPE_CHECKING:
@@ -69,13 +71,18 @@ class Bot:
 
         return self.__token
 
-    def set_name(self, name: str):
+    @property
+    def name(self) -> str:
+        """Bot's username. Can be set one time."""
+
+        return self.__name
+
+    @name.setter
+    def name(self, name: str):
         if self.__name is not None:
             raise TypeError('Bot''s name can be set one time.')
 
         self.__name = name
-
-    name = property(fget=lambda self: self.__name, fset=set_name, doc="Bot's username. Can be set one time.")
 
     @property
     def user_id(self) -> int:
@@ -108,7 +115,7 @@ class Bot:
 
         logger.debug('Performing init...')
 
-        context.current_bot.set(self)
+        context2.bot = self
 
         if self.__own_connector:
             await self.connector.init()
@@ -129,7 +136,7 @@ class Bot:
 
         logger.debug('Performing shutdown...')
 
-        context.current_bot.set(self)
+        context2.bot = self
 
         await self.router.shutdown()
 
@@ -144,44 +151,45 @@ class Bot:
     async def process(self, executor: 'Executor', update: Update) -> Optional[Request]:
         logger_raw_in.debug('Raw in: %s', update.raw)
 
-        context.current_executor.set(executor)
-        context.current_bot.set(self)
-        context.current_webhook_requests.set(list())
+        context2.executor = executor
+        context2.bot = self
 
-        context.current_update.set(update)
-        context.current_message.set(None)
-        context.current_chat.set(None)
-        context.current_user.set(None)
+        context2.webhook_requests = list()
+
+        context2.update = update
+        context2.message = None
+        context2.chat = None
+        context2.user = None
 
         if update.update_type is UpdateType.message:
-            context.current_message.set(update.message)
-            context.current_chat.set(update.message.chat)
-            context.current_user.set(update.message.user)
+            context2.message = update.message
+            context2.chat = update.message.chat
+            context2.user = update.message.user
         elif update.update_type is UpdateType.edited_message:
-            context.current_message.set(update.edited_message)
-            context.current_chat.set(update.edited_message.chat)
-            context.current_user.set(update.edited_message.user)
+            context2.message = update.edited_message
+            context2.chat = update.edited_message.chat
+            context2.user = update.edited_message.user
         elif update.update_type is UpdateType.channel_post:
-            context.current_message.set(update.channel_post)
-            context.current_chat.set(update.channel_post.chat)
-            context.current_user.set(update.channel_post.user)
+            context2.message = update.channel_post
+            context2.chat = update.channel_post.chat
+            context2.user = update.channel_post.user
         elif update.update_type is UpdateType.edited_channel_post:
-            context.current_message.set(update.edited_channel_post)
-            context.current_chat.set(update.edited_channel_post.chat)
-            context.current_user.set(update.edited_channel_post.user)
+            context2.message = update.edited_channel_post
+            context2.chat = update.edited_channel_post.chat
+            context2.user = update.edited_channel_post.user
         elif update.update_type is UpdateType.inline_query:
-            context.current_user.set(update.inline_query.user)
+            context2.user = update.inline_query.user
         elif update.update_type is UpdateType.chosen_inline_result:
-            context.current_user.set(update.chosen_inline_result.user)
+            context2.user = update.chosen_inline_result.user
         elif update.update_type is UpdateType.callback_query:
-            context.current_message.set(update.callback_query.message)
+            context2.message = update.callback_query.message
             if update.callback_query.message:
-                context.current_chat.set(update.callback_query.message.chat)
-            context.current_user.set(update.callback_query.user)
+                context2.chat = update.callback_query.message.chat
+            context2.user = update.callback_query.user
         elif update.update_type is UpdateType.shipping_query:
-            context.current_user.set(update.shipping_query.user)
+            context2.user = update.shipping_query.user
         elif update.update_type is UpdateType.pre_checkout_query:
-            context.current_user.set(update.pre_checkout_query.user)
+            context2.user = update.pre_checkout_query.user
 
         try:
             for md in self.__middlewares:
@@ -193,7 +201,7 @@ class Bot:
 
             webhook_request = None
 
-            for req in context.get_webhook_requests():
+            for req in context2.webhook_requests:
                 # set request to return if it can be processed
                 if webhook_request is None and executor.can_process_webhook_request(req):
                     for md in self.__middlewares: