+++ rocketgram/routers/dispatcher/commonwaiters.py
@@ -5,7 +5,7 @@
 
 from . import commonfilters
 from .waiters import make_waiter, DropWaiter
-from ... import context
+from ...context import context2
 from ...update import UpdateType, MessageType
 
 
@@ -22,7 +22,7 @@ def next_message(*message_types: MessageType):
     if not message_types:
         message_types = (MessageType.text,)
 
-    if context.message().message_type in message_types:
+    if context2.message.message_type in message_types:
         return True
 
     return False